Moved everything into a different class
