+++ src/main/java/TaskList.java
@@ -1,0 +1,10 @@
+import java.util.ArrayList;
+import java.util.List;
+
+public class TaskList {
+    static List<ListItem> listItems = new ArrayList<ListItem>();
+
+    public static List<ListItem> getListItems() {
+        return listItems;
+    }
+}
